+++ R/validate_linelist.R
@@ -22,7 +22,8 @@
 #'
 #' @inheritParams set_tags
 #'
-#' @return If checks pass, a `linelist` object; otherwise issues an error.
+#' @return If checks pass, a `linelist` object (invisibly); otherwise issues an
+#' error.
 #'
 #' @seealso
 #' * [tags_types()] to change allowed types
@@ -66,5 +67,7 @@ validate_linelist <- function(x,
   validate_tags(x, allow_extra)
   validate_types(x, ref_types)
 
-  x
+  message("'", checkmate::vname(x), "' is a valid linelist object")
+
+  invisible(x)
 }